+++ typescript/simple-mercari-web/src/App.tsx
@@ -1,11 +1,10 @@
 //@ts-ignore
 import { useState } from 'react';
-import { Route, useHistory, Link, Switch, BrowserRouter as Router} from 'react-router-dom';
+import { Route, useHistory, Switch, BrowserRouter as Router} from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 import { ItemList } from './components/ItemList';
-
-import { Sell } from './components/Sell'
+import { Listing } from './components/Listing';
 
 function App() {
   // reload ItemList after Listing complete
@@ -13,10 +12,11 @@ function App() {
 
   const history=useHistory();
 
-  const onClickGoSell=()=>history.push("/sell")
+  const onClickGoSell=()=>history.goback()
 
   return (
     <div>
+      
       <Router>
         
         <Switch>
@@ -24,16 +24,13 @@ function App() {
           <div>
             <ItemList reload={reload} onLoadCompleted={() => setReload(false)} />
             <button onClick={onClickGoSell}>出品</button>
+            <p>商品一覧表示画面</p>
           </div>
         </Route>
         <Route path="/sell">
-          <Sell setReload={setReload} />
+          <Listing onListingCompleted={() => setReload(true)} />
         </Route>
         </Switch>
-        
-        {/* <Route path="/sell"> */}
-          
-        {/* </Route> */}
       </Router>
       
     </div>
